api_gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
bcrypt==4.1.2
python-dateutil==2.8.2
requests==2.31.0
celery==5.3.4
redis==5.0.1